+++ src/PimEnterprise/Bundle/WorkflowBundle/Model/PublishedProductMediaInterface.php
@@ -5,6 +5,9 @@ namespace PimEnterprise\Bundle\WorkflowBundle\Model;
 /**
  * Published product media interface
  *
+ * TODO: to remove
+ * @deprecated will be removed in 1.4
+ *
  * @author Julien Janvier <julien.janvier@akeneo.com>
  */
 interface PublishedProductMediaInterface